need to refactor to account for imu not loop around fifo address - need to send fifo read address with every read transaction

+++ src/experimentation/combo2/light_sensor.h
@@ -46,8 +46,8 @@ public:
     uint32_t getBrightness() const;
 
     // IMultiDmaTransactionSource Interface
-    int getRequiredDescriptorCount(uint32_t frame_id, uint8_t subframe_id, uint8_t subframe_max) override;
-    void populateDescriptors(uint32_t frame_id, uint8_t subframe_id, uint8_t subframe_max, DmaDescriptor* pool_start, int data_channel, int aux0_channel, int aux1_channel, int ctrl_channel) override;
+    int getRequiredDescriptorCount(uint64_t frame_id, uint8_t subframe_id, uint8_t subframe_max) override;
+    void populateDescriptors(uint64_t frame_id, uint8_t subframe_id, uint8_t subframe_max, DmaDescriptor* pool_start, int data_channel, int aux0_channel, int aux1_channel, int ctrl_channel) override;
 };
 
 #endif
